+++ Crypto/LoadMarks.cs
@@ -159,12 +159,12 @@ namespace Priem
                     newRow["ФИО"] = pm.FIO;
                     newRow["Номер"] = pm.PersonVedNumber;
                     lstNumbers.Add(pm.PersonVedNumber);
-                    
-                    int? sumMark = 0;
+
+                    decimal? sumMark = 0;
                     if(pm.Mark == null && pm.OralMark == null)
                         sumMark = null;
                     else                    
-                        sumMark = (pm.Mark ?? 0) + (pm.OralMark ?? 0); 
+                        sumMark = (pm.Mark ?? 0m) + (pm.OralMark ?? 0m); 
                   
                     newRow["Баллы"] = sumMark;
                     examTable.Rows.Add(newRow);
@@ -305,11 +305,11 @@ namespace Priem
                         td[8, i] = examId;
                         td[9, i] = examDate.ToShortDateString();
 
-                        int? sumMark = 0;
+                        decimal? sumMark = 0;
                         if (pm.Mark == null && pm.OralMark == null)
                             sumMark = null;
                         else
-                            sumMark = (pm.Mark ?? 0) + (pm.OralMark ?? 0);
+                            sumMark = (pm.Mark ?? 0m) + (pm.OralMark ?? 0m);
                                           
                         td[10, i] =  sumMark == null ? "" : sumMark.ToString();
